Move more submission components
